Move multi-instance parser elements inside specific package
