+++ activiti-core/activiti-bpmn-converter/src/main/java/org/activiti/bpmn/converter/child/multi/instance/LoopDataOutputRefParser.java
@@ -14,24 +14,25 @@
  * limitations under the License.
  */
 
-package org.activiti.bpmn.converter.child;
+package org.activiti.bpmn.converter.child.multi.instance;
 
-import static org.activiti.bpmn.constants.BpmnXMLConstants.ELEMENT_MULTIINSTANCE_CONDITION;
+import static org.activiti.bpmn.constants.BpmnXMLConstants.ELEMENT_MULTI_INSTANCE_DATA_OUTPUT;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamReader;
+import org.activiti.bpmn.converter.child.ElementParser;
 import org.activiti.bpmn.model.MultiInstanceLoopCharacteristics;
 
-public class MultiInstanceCompletionConditionParser implements ElementParser<MultiInstanceLoopCharacteristics> {
+public class LoopDataOutputRefParser implements ElementParser<MultiInstanceLoopCharacteristics> {
 
     @Override
     public boolean canParseCurrentElement(XMLStreamReader reader) {
-        return reader.isStartElement() && ELEMENT_MULTIINSTANCE_CONDITION.equalsIgnoreCase(reader.getLocalName());
+        return reader.isStartElement() && ELEMENT_MULTI_INSTANCE_DATA_OUTPUT.equalsIgnoreCase(reader.getLocalName());
     }
 
     @Override
     public void setInformation(XMLStreamReader reader,
         MultiInstanceLoopCharacteristics loopCharacteristics) throws XMLStreamException {
-        loopCharacteristics.setCompletionCondition(reader.getElementText());
+        loopCharacteristics.setLoopDataOutputRef(reader.getElementText());
     }
 }
