+++ activiti-core/activiti-bpmn-converter/src/main/java/org/activiti/bpmn/converter/child/multi/instance/MessageEventDefinitionParser.java
@@ -1,22 +1,27 @@
-/* Licensed under the Apache License, Version 2.0 (the "License");
+/*
+ * Copyright 2019 Alfresco, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.bpmn.converter.child;
+
+package org.activiti.bpmn.converter.child.multi.instance;
 
 import java.util.Collections;
 import java.util.List;
 
 import javax.xml.stream.XMLStreamReader;
 
+import org.activiti.bpmn.converter.child.BaseChildElementParser;
 import org.activiti.bpmn.converter.util.BpmnXMLUtil;
 import org.activiti.bpmn.model.BaseElement;
 import org.activiti.bpmn.model.BpmnModel;
@@ -25,9 +30,6 @@ import org.activiti.bpmn.model.ExtensionAttribute;
 import org.activiti.bpmn.model.MessageEventDefinition;
 import org.apache.commons.lang3.StringUtils;
 
-/**
-
- */
 public class MessageEventDefinitionParser extends BaseChildElementParser {
 
   public String getElementName() {
